+++ .github/workflows/publish-release.yml
@@ -9,7 +9,6 @@ on:
 env:
   DOCKERHUB_REGISTRY: docker.io
   GITHUB_REGISTRY: ghcr.io
-  DOCKERHUB_IMAGE_NAME: ${{ secrets.DOCKER_USERNAME }}/${{ github.event.repository.name }}
   GITHUB_IMAGE_NAME: ${{ github.repository }}
 
 concurrency:
@@ -69,7 +68,22 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v3.8.0
 
+      - name: Resolve registry images
+        id: registries
+        env:
+          DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
+        run: |
+          {
+            echo 'images<<EOF'
+            if [ -n "$DOCKER_USERNAME" ]; then
+              echo "${DOCKER_USERNAME}/${{ github.event.repository.name }}"
+            fi
+            echo "${{ env.GITHUB_REGISTRY }}/${{ env.GITHUB_IMAGE_NAME }}"
+            echo EOF
+          } >> "$GITHUB_OUTPUT"
+
       - name: Log in to Docker Hub
+        if: ${{ secrets.DOCKER_USERNAME != '' && secrets.DOCKER_PASSWORD != '' }}
         uses: docker/login-action@v3.3.0
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
@@ -119,9 +133,7 @@ jobs:
         id: meta
         uses: docker/metadata-action@v5.6.1
         with:
-          images: |
-            ${{ env.DOCKERHUB_IMAGE_NAME }}
-            ${{ env.GITHUB_REGISTRY }}/${{ env.GITHUB_IMAGE_NAME }}
+          images: ${{ steps.registries.outputs.images }}
           tags: |
             type=raw,value=${{ steps.vars.outputs.latest_tag }}
             type=raw,value=${{ steps.vars.outputs.release_tag }}
